requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
openai>=1.30.0
ddtrace>=2.9.0
datadog>=0.48.0
Pillow>=10.3.0
